Finalize plans
